refactor: TableGroupValidator 의 위치를 application 으로 변경하라
