+++ src/main/java/kitchenpos/tableGroup/application/TableGroupValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.tableGroup.domain;
+package kitchenpos.tableGroup.application;
 
 import java.util.List;
 import kitchenpos.common.exception.BadRequestException;
